initialized pointer vars with =>NULL() which are passed in calls to routine storage_get_levels() in metgrid/storage_module.F
The allocated(pointer_var) call fails (runtime failure) using the intel 2017.035 compiler when the pointer variable is not defined, and this will always happen.  Apparently, the default Gnu compiler behaviour is for allocated() to return .FALSE. when either undefined or NULL, and this behavior is expected even if outside the Fortran standard.